# feature/optimize-tiff-reader-and-add-benchmarks