--- a/.gitignore
+++ b/.gitignore
@@ -121,3 +121,4 @@ resources/
 # asv / benchmarks
 .asv/html
 .asv/env
+.asv/results