… .github/workflows/build-docs.yml
… .github/workflows/test-and-lint.yml
… CONTRIBUTING.md
… README.md
… aicsimageio/tests/readers/reader_test_utils.py
… aicsimageio/tests/readers/test_tiff_reader.py
… asv.conf.json
… benchmarks/__init__.py
… benchmarks/benchmark_lib.py
… benchmarks/benchmark_readers.py
… benchmarks/benchmarks.py
… docs/benchmarks.md
… docs/index.rst
… scripts/benchmark.py
… scripts/chart_benchmarks.py