## Description
This reduces the chunked read time for TIFFs (and OME-TIFF) by ~92% from what it is now.
(for "real" files, improvement ranges from an 87% decrease to a 96% decrease in chunked read time)
Much of the conversation and tracking for this work can be found in #178.

This PR points at the 4.x development head `feature/update-to-4.0-api` rather than `master`. To this point, this "resolves" #178 for 4.x but not for 3.x. I am unsure if I will get around to backporting the optimizations.

Here is the result of running `asv continuous feature/update-to-4.0-api feature/benchmarks`:
```
       before           after         ratio
     [ab87cd1a]       [a3d3e970]
     <feature/update-to-4.0-api>       <feature/benchmarks>
+      28.1±0.9ms         36.7±2ms     1.30  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_10_c_3_z_1.tiff')
+      28.1±0.2ms         36.4±1ms     1.30  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_10_c_3_z_1.tiff')
-             96M            87.1M     0.91  benchmark_readers.TiffReaderSuite.peakmem_delayed_array('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-            101M            84.3M     0.84  benchmark_readers.TiffReaderSuite.peakmem_delayed_array('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-         210±3ms        172±0.4ms     0.82  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-        58.8±2ms         47.2±1ms     0.80  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-        56.7±1ms         44.9±1ms     0.79  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-        74.9±4ms         58.6±2ms     0.78  benchmark_readers.OmeTiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-        53.3±3ms       40.0±0.5ms     0.75  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1.ome.tiff')
-        622±10ms          450±1ms     0.72  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pipeline-4.ome.tiff')
-         130±2ms         93.7±1ms     0.72  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-      1.07±0.01s          754±7ms     0.70  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pre-variance-cfe.ome.tiff')
-      1.05±0.01s          731±6ms     0.70  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-      36.6±0.7ms       25.1±0.7ms     0.68  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1.ome.tiff')
-      1.03±0.01s          689±6ms     0.67  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-         414±5ms          253±2ms     0.61  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-         441±9ms          249±1ms     0.56  benchmark_readers.OmeTiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-      3.52±0.01s       1.83±0.01s     0.52  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1_RGB.tiff')
-      3.51±0.01s       1.81±0.01s     0.51  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1_RGB.tiff')
-         362±3ms          181±2ms     0.50  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/actk.ome.tiff')
-      52.3±0.9ms       24.9±0.4ms     0.48  benchmark_readers.OmeTiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-         348±7ms          163±1ms     0.47  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-        565±10ms          246±5ms     0.43  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_3_t_1_c_3_z_5.ome.tiff')
-         302±5ms          129±1ms     0.43  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pipeline-4.ome.tiff')
-         488±2ms          180±2ms     0.37  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-         456±6ms          158±1ms     0.35  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pre-variance-cfe.ome.tiff')
-      31.1±0.8ms      8.43±0.05ms     0.27  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_10_z_1.ome.tiff')
-         252±5ms       68.2±0.1ms     0.27  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/actk.ome.tiff')
-         455±5ms        113±0.7ms     0.25  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-      3.19±0.02s          428±2ms     0.13  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/actk.ome.tiff')
-      6.08±0.01s         789±10ms     0.13  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-      6.20±0.03s         797±20ms     0.13  benchmark_readers.OmeTiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-      6.54±0.01s          841±5ms     0.13  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pre-variance-cfe.ome.tiff')
-         7.00±0s         839±20ms     0.12  benchmark_readers.OmeTiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-      6.90±0.02s          827±8ms     0.12  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-            768M            83.4M     0.11  benchmark_readers.TiffReaderSuite.peakmem_delayed_array('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1_RGB.tiff')
-      3.08±0.04s          312±2ms     0.10  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/actk.ome.tiff')
-      5.95±0.03s          593±4ms     0.10  benchmark_readers.OmeTiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pipeline-4.ome.tiff')
-      14.9±0.06s       1.42±0.01s     0.10  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/actk.ome.tiff')
-      5.62±0.03s          260±5ms     0.05  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pipeline-4.ome.tiff')
-      5.73±0.03s          255±5ms     0.04  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/pre-variance-cfe.ome.tiff')
-      6.31±0.02s          281±3ms     0.04  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-      6.40±0.02s          284±2ms     0.04  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/variance-cfe.ome.tiff')
-         5.43±0s          213±2ms     0.04  benchmark_readers.TiffReaderSuite.time_random_single_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-      5.46±0.02s          208±2ms     0.04  benchmark_readers.TiffReaderSuite.time_random_many_chunk_read('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/3d-cell-viewer.ome.tiff')
-      1.72±0.01s      1.72±0.05ms     0.00  benchmark_readers.TiffReaderSuite.time_delayed_array_construct('/home/maxfield/active/cell/aicsimageio/aicsimageio/tests/resources/s_1_t_1_c_1_z_1_RGB.tiff')
```

These results are using the default `chunk_by_dims`, in many cases users should tailor the `chunk_by_dims` to their file. I.E. a timeseries file should probably `chunk_by_dims=["T", "Y", "X"]`

Also in this PR:
* You can now call `items` on the `Dimensions` object: `for dim, size in reader.dims.items()`.
* Removal of old `scripts/benchmarks.py` and `scripts/chart-benchmarks.py`.
* We no longer need to use a specific quilt hash for benchmark files because some of our test files are large anyway, but we could if we wanted to.
* A bunch of small text changes for documentation and other files that happened as a part of formatting / auto-save.

## Pull request recommendations:
- [x] Name your pull request _your-development-type/short-description_. Ex: _feature/read-tiff-files_
- [x] Link to any relevant issue in the PR description. Ex: _Resolves [gh-12], adds tiff file format support_

Resolves-for-4.x #178 
Resolves #105 

- [x] Provide relevant tests for your feature or bug fix.
- [x] Provide or update documentation for any feature added by your pull request.

I have a working version of this PR on my fork.
You can see benchmarks working per commit [here](https://jacksonmaxfield.github.io/aicsimageio/_benchmarks/index.html#/). (That is using a fake regression and fix to show that it works)
Benchmarks are ran as a part of the documentation generation workflow. Each commit's results are stored in the [`benchmark-results` branch](https://github.com/JacksonMaxfield/aicsimageio/tree/benchmark-results).

Thanks for contributing!