@JacksonMaxfield 2021-01-21
> Not sure if I saw readme instructions for running the benchmarks locally outside of gh-actions.

Added a section to CONTRIBUTING about running benchmarks on patches that touch a base reader.